Consolidated platform dialogs into a folder
